requirements.txt
langchain>=0.3.0
langchain-google-genai>=3.1.0
langgraph>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pillow>=10.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.31.0
